Add modal dialog to confirm role removal when remove role btn is clicked

+++ MBDB-app/Scripts/ManageRoles/functions/RemoveRoleFromUser.js
@@ -7,21 +7,27 @@
         roleName
     }
 
-    $.ajax({
-        url: "/api/manageusers",
-        data: data,
-        method: "DELETE"
+    bootbox.confirm("Are you sure you want to remove " + roleName + " from this user?", (result) => {
+        if (result) {
+            $.ajax({
+                url: "/api/manageusers",
+                data: data,
+                method: "DELETE"
+            })
+                .done((data) => {
+                    toastr.success("Successfully removed the role from this user!");
+                    $(userRoles).html("");
+                    let unorderedList = createListOfRoles(data, userId);
+
+                    $(userRoles).append(unorderedList);
+
+                    updateRolesInTheTable(userId);
+                })
+                .fail((data) => {
+                    toastr.error(data.responseJSON.message)
+                })
+        }
     })
-        .done((data) => {
-            toastr.success("Successfully removed the role from this user!");
-            $(userRoles).html("");
-            let unorderedList = createListOfRoles(data, userId);
 
-            $(userRoles).append(unorderedList);
-
-            updateRolesInTheTable(userId);
-        })
-        .fail((data) => {
-            toastr.error(data.responseJSON.message)
-        })
+    
 }
